Refactor on `export` - define `index.ts` from `projects/core` => import { Project, Task } from '@src/projects/core' - define `index.ts` from `projects/use-cases` - from every query => request export itself within payload & response - from every command => request export itself within payload
