+++ src/projects/core/index.ts
@@ -1,0 +1,2 @@
+export * from "./project";
+export * from "./task";
