+++ src/projects/persistence/schemas/project.schema.ts
@@ -1,8 +1,7 @@
 import { EntitySchema } from "typeorm";
 import { Project } from "@projects/core/project";
 import { EntityBaseSchema } from "@building-blocks/infra/database/schemas/entity-base-schema";
-import { Task } from "@src/projects/core/task";
-import { TaskSchema } from "./task.schema";
+import { Task } from "@src/projects/core";
 
 export const ProjectSchema = new EntitySchema<Project>({
     name: Project.name,
