+++ src/projects/persistence/schemas/task.schema.ts
@@ -1,7 +1,7 @@
 import { Task } from "@projects/core/task";
 import { EntityBaseSchema } from "@building-blocks/infra/database/schemas/entity-base-schema";
 import { EntitySchema, EntitySchemaColumnOptions } from "typeorm";
-import { Project } from "@src/projects/core/project";
+import { Project } from "@src/projects/core";
 import { ProjectSchema } from "./project.schema";
 
 const columns = {
