+++ src/projects/projects.module.ts
@@ -1,5 +1,5 @@
 import { Global, Module } from '@nestjs/common';
-import { CommandHandlers, QueryHandlers } from './use-cases';
+import { Handlers } from './use-cases';
 import { ProjectsController } from './projects.controller';
 import { CqrsModule } from '@nestjs/cqrs';
 import { TypeOrmModule } from '@nestjs/typeorm';
@@ -11,10 +11,7 @@ import { ProjectsModuleDataSource } from './persistence';
         CqrsModule, 
         TypeOrmModule.forFeature([...ProjectsModuleDataSource.Schemas])
     ],
-    providers: [
-        ...CommandHandlers,
-        ...QueryHandlers
-    ],
+    providers: [ ...Handlers ],
     controllers: [ProjectsController]
 })
