+++ src/projects/use-cases/index.ts
@@ -1,13 +1,19 @@
-import { CreateProjectHandler } from "./project/commands/create/create-project.handler";
+
 import { FindByIdHandler } from "./project/queries/find-by-id/find-by-id.handler";
 import { SearchProjectsHandler } from "./project/queries/search/search-projects.handlers";
+import { CreateProjectHandler } from "./project/commands/create/create-project.handler";
 
-
-export const CommandHandlers = [
+const CommandHandlers = [
     CreateProjectHandler
 ]
 
-export const QueryHandlers = [
+const QueryHandlers = [
     SearchProjectsHandler,
     FindByIdHandler
 ]
+
+export const Handlers = [ ...CommandHandlers, ...QueryHandlers ];
+
+export * from "./project/commands/create/create-project.request";
+export * from "./project/queries/find-by-id/find-by-id.request";
+export * from "./project/queries/search/search-projects.request";
