+++ src/projects/use-cases/project/commands/create/create-project.request.ts
@@ -1,5 +1,6 @@
 import { CreateProjectPayload } from "./create-project.payload";
 
+export * from "./create-project.payload";
 
 export class CreateProjectRequest {
     constructor(
