+++ src/projects/use-cases/project/queries/find-by-id/find-by-id.handler.ts
@@ -2,9 +2,9 @@ import { IQueryHandler, QueryHandler } from "@nestjs/cqrs";
 import { FindByIdRequest } from "./find-by-id.request";
 import { FindByIdResponse } from "./find-by-id.response";
 import { InjectRepository } from "@nestjs/typeorm";
-import { Project } from "@projects/core/project";
 import { Repository } from "typeorm";
 import { CachingProvider } from "@src/building-blocks/infra/caching/caching.provider";
+import { Project } from "@src/projects/core";
 
 @QueryHandler(FindByIdRequest)
 export class FindByIdHandler implements IQueryHandler<FindByIdRequest, FindByIdResponse> {
