+++ src/projects/use-cases/project/queries/find-by-id/find-by-id.request.ts
@@ -1,5 +1,7 @@
 import { IQuery } from "@nestjs/cqrs";
 
+export * from "./find-by-id.response";
+
 export class FindByIdRequest implements IQuery{
     constructor(public readonly id: string) {}
 }
