+++ src/projects/use-cases/project/queries/find-by-id/find-by-id.response.ts
@@ -1,8 +1,6 @@
 import { IQueryResult } from "@nestjs/cqrs";
 import { ApiProperty } from "@nestjs/swagger";
-import { Project } from "@src/projects/core/project";
-import { Task } from "@src/projects/core/task";
-
+import { Project, Task } from "@src/projects/core";
 
 class TaskDto {
     @ApiProperty({ type: String })
