+++ test/projects/search.e2e-spec.ts
@@ -4,8 +4,8 @@ import { Repository } from 'typeorm';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import { Project } from '@src/projects/core/project';
 import { app, httpServer } from '@test/test.setup';
-import { faker, sk } from '@faker-js/faker';
-import { ProjectDto } from '@src/projects/use-cases/project/queries/search/search-projects.response';
+import { faker } from '@faker-js/faker';
+import { ProjectDto } from '@src/projects/use-cases';
 
 const numberOfProjects = 4;
 const testCases = [
